cont'd work on Tiled parsing for tileset - created Tile class

+++ nova2d/nova-zero/src/utils/timeline/Timeline.cpp
@@ -28,7 +28,7 @@ namespace novazero
 					{
 						m_TimelineEvents[m_CurrentEvent]->Execute();
 
-						if (m_TimelineEvents.size() > m_CurrentEvent + 1)
+						if (m_TimelineEvents.size() > (size_t)m_CurrentEvent + 1)
 						{
 							m_TimelineEvents[m_CurrentEvent]->SetRunning(false);
 							m_CurrentEvent++;
@@ -60,7 +60,7 @@ namespace novazero
 
 			void Timeline::SetEvent(int index)
 			{
-				if (index >= 0 && index < m_TimelineEvents.size())
+				if (index >= 0 && index < (int)m_TimelineEvents.size())
 				{
 					m_CurrentEvent = index;
 				}
